+++ cpp/core/shuffle/Partitioner.h
@@ -37,7 +37,7 @@ class Partitioner {
   virtual arrow::Status compute(
       const int32_t* pidArr,
       const int64_t numRows,
-      std::vector<uint16_t>& row2partition,
+      std::vector<uint32_t>& row2partition,
       std::vector<uint32_t>& partition2RowCount) = 0;
 
  protected:
